// packages/client/vms/interfaces/IBlockVM.ts
export interface IBlockVM {
    readonly id: string
    readonly type: string
    properties: string
    readonly content: string[]
    readonly parent_id: string;
    readonly space_id: string;
    readonly last_updated: Date;
    last_modified: Date;
    readonly created_time: Date
    readonly created_by_id: string
}


export interface IBlockTreeVM {
    value:{
        [key: string]: IBlockVM

    }
}